+++ flamedisx/nest/lxe_blocks/detection.py
@@ -136,52 +136,3 @@ class DetectElectrons(DetectPhotonsOrElectrons):
         return super()._compute(quanta_produced=electrons_produced,
                                 quanta_detected=electrons_detected,
                                 data_tensor=data_tensor, ptensor=ptensor)
-
-
-@export
-class DetectS2Photons(fd.Block):
-    dimensions = ('electrons_detected', 's2_photons_produced')
-    extra_dimensions = ()
-
-    model_functions = ('electron_gain_mean', 'electron_gain_std')
-
-    def _compute(self, data_tensor, ptensor,
-                 electrons_detected, s2_photons_produced):
-        mean_per_q = self.gimme('electron_gain_mean',
-                                data_tensor=data_tensor,
-                                ptensor=ptensor)[:, o, o]
-        std_per_q = self.gimme('electron_gain_std',
-                               data_tensor=data_tensor,
-                               ptensor=ptensor)[:, o, o]
-
-        mean = electrons_detected * mean_per_q
-        std = electrons_detected ** 0.5 * std_per_q
-
-        # add offset to std to avoid NaNs from norm.pdf if std = 0
-        result = tfp.distributions.Normal(
-            loc=mean, scale=std + 1e-10
-        ).prob(s2_photons_produced)
-
-        return result
-
-    def _simulate(self, d):
-        d['s2_photons_produced'] = stats.norm.rvs(
-            loc=(d['electrons_detected']
-                 * self.gimme_numpy('electron_gain_mean')),
-            scale=(d['electrons_detected']**0.5
-                   * self.gimme_numpy('electron_gain_std')))
-
-    def _annotate(self, d):
-        m = self.gimme_numpy('electron_gain_mean')
-        s = self.gimme_numpy('electron_gain_std')
-
-        mle = d['electrons_detected_mle'] = \
-            (d['s2_photons_produced_mle'] / m).clip(0, None)
-
-        scale = mle**0.5 * s / m
-
-        for bound, sign, intify in (('min', -1, np.floor),
-                                    ('max', +1, np.ceil)):
-            d['electrons_detected_' + bound] = intify(
-                mle + sign * self.source.max_sigma * scale
-            ).clip(0, None).astype(np.int)
